Migrate from Packer.nvim to Lazy.nvim
