+++ .gitignore
@@ -11,9 +11,13 @@ runcom/vim/bundle/*
 runcom/vim/plugged/*
 runcom/vim/vim
 runcom/vim/spell/*
+runcom/vim/undodir/*
 
 **/spell/**
 
 #---------------------------------VScode--------------------------#
 .vscode/
 .lhistory
+
+#---------------------------------NeoVim--------------------------#
+runcom/config/nvim/lazy-lock.json
