+++ runcom/config/nvim/lua/plugins/playground.lua
@@ -1,0 +1,4 @@
+return {
+    'nvim-treesitter/playground',
+    event= 'VeryLazy'
+}
