+++ runcom/config/nvim/lua/plugins/rosepine.lua
@@ -1,0 +1,7 @@
+return {
+    'rose-pine/neovim',
+    name = 'rose-pine',
+    config = function()
+        vim.cmd('colorscheme rose-pine')
+    end
+}
