+++ runcom/config/nvim/lua/plugins/telescope.lua
@@ -1,0 +1,13 @@
+return {
+    'nvim-telescope/telescope.nvim', version= '0.1.6',
+    -- or                            , branch = '0.1.x',
+    dependencies = { {'nvim-lua/plenary.nvim'} },
+    config = function()
+        local builtin = require('telescope.builtin')
+        vim.keymap.set('n', '<leader>pf', builtin.find_files, {})
+        vim.keymap.set('n', '<C-p>', builtin.git_files, {})
+        vim.keymap.set('n', '<leader>ps', function()
+            builtin.grep_string({ search = vim.fn.input("Grap > ") });
+        end)
+    end
+}
