+++ runcom/config/nvim/lua/plugins/treesitter.lua
@@ -1,0 +1,5 @@
+return { 
+    'nvim-treesitter/nvim-treesitter', 
+    build = ':TSUpdate',
+    event= 'VeryLazy',
+}
