+++ runcom/config/nvim/lua/plugins/undotree.lua
@@ -1,0 +1,8 @@
+return {
+    'mbbill/undotree',
+    event= 'VeryLazy',
+    config = function()
+        vim.keymap.set('n', '<leader>u', vim.cmd.UndotreeToggle)
+    end
+
+}
