+++ runcom/config/nvim/lua/swarmies/init.lua
@@ -1,6 +1,4 @@
 require("swarmies.remap")
 require("swarmies.set")
-require("swarmies.packer")
--- require("swarmies.lazy")
 print("hello from swarmies")
